+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - wapm will now suggest a package to install that contains the desired command if the command is not found
 
 ### Changed
+- Renamed Wasm Contracts to Wasm Interfaces
 - Lockfile version 3 with package root directory added
 
 ## [0.3.1] - 2019-06-19